Some more security provided

+++ pages/api/products/[id].js
@@ -6,8 +6,11 @@ import { Cagliostro } from '@next/font/google';
 export default async function handler(req, res) {
     const {
         method,
-        query: { id }
+        query: { id },
+        cookies,
     } = req;
+
+    const token = cookies.token;
 
     await dbConnect();
 
@@ -20,16 +23,10 @@ export default async function handler(req, res) {
         }
     }
 
-    if (method === 'POST') {
-        try {
-            const product = await Product.create(req.body);
-            res.status(201).json(product);
-        }catch(err) {
-            res.status(500).json(err);
-        }
-    }
-
     if (method === 'PUT') {
+        if (!token || token !== process.env.TOKEN) {
+            return res.status(401).json("Not authenticated!");
+        }
         try {
             const product = await Product.findByIdAndUpdate(id, req.body, {
                 new: true
@@ -41,6 +38,9 @@ export default async function handler(req, res) {
     }
 
     if (method === 'DELETE') {
+        if (!token || token !== process.env.TOKEN) {
+            return res.status(401).json("Not authenticated!");
+        }
         try {
             await Product.findByIdAndDelete(id);
             res.status(201).json("The product has been deleted!");
